fix bug

+++ src/main/java/sokhoahoccongnghe/phutho/gov/vn/service/implement/UploadAndDownloadServiceImpl.java
@@ -104,10 +104,12 @@ public class UploadAndDownloadServiceImpl implements UploadAndDownloadService {
     @Transactional
     public void deleteForm(Integer formId) {
         String deletedFileName = formService.deleteById(formId);
-        try {
-            fileStorageService.deleteFile(deletedFileName, FileTypeEnum.FORM.getValue());
-        } catch (IOException e) {
-            throw new FileDeleteException("can not delete file", e);
+        if(deletedFileName != null) {
+            try {
+                fileStorageService.deleteFile(deletedFileName, FileTypeEnum.FORM.getValue());
+            } catch (IOException e) {
+                throw new FileDeleteException("can not delete file", e);
+            }
         }
     }
